Refactor redux folder and authentication slice
