+++ src/views/App.js
@@ -6,7 +6,7 @@ import {
 } from 'react-router-dom';
 import { Provider } from 'react-redux';
 
-import store from './shared/redux';
+import store from './../redux';
 
 // Pages
 import Home from './home/Home';
