+++ src/views/checkout/CheckoutLoader.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useParams } from 'react-router-dom';
 import { useSelector } from 'react-redux';
-import { selectJwt } from '../shared/redux/authorization';
+import { selectJwt } from '../../redux/authenticationSlice';
 import Loader from '../shared/components/Loader';
 import { getCartByUsername } from '../../api/checkoutApi';
 import Checkout from './Checkout';
