+++ src/views/loginCallback/LoginCallbackHandler.js
@@ -1,13 +1,13 @@
 import { useHistory, useLocation } from 'react-router-dom';
 import { useDispatch } from 'react-redux';
 import jwtDecode from 'jwt-decode';
-import { login, setUserData } from '../shared/redux/authorization';
+import { login, setUserData } from '../../redux/authenticationSlice';
 import getHashParams from '../shared/helpers/getHashParams';
 import { getUser } from '../../api/usersApi';
 
 // Receives a cognito code, and uses it to make a request for a JWT.
 //     Then, stores JWT in Redux state
-const LoginCallBackHandler = props => {
+const LoginCallBackHandler = () => {
   const history = useHistory();
   const location = useLocation();
   const dispatch = useDispatch();
@@ -18,13 +18,15 @@ const LoginCallBackHandler = props => {
   const jwt = hashObj.id_token;
 
   if (jwt) {
-    // Decode and save jwt
+    // Decode and save jwt (account data) to redux
     const accountData = jwtDecode(jwt);
     const payload = { jwt, accountData };
     dispatch(login(payload));
 
     // Fetch user data asynchronously. When done, save it to redux.
     const username = accountData['cognito:username'];
+    // TODO user may still be in the process of creating
+    //     Fix: if 404 throw and use exponential backup retry
     getUser(username).then(userAndOrders => dispatch(setUserData(userAndOrders.user)));
 
     history.push('/?message=loginSuccess');
